Fixing compiler warnings seen on Mac builds
1. Bring fraction.h into sync with master, as far as possible. Fixes a lot of warnings reg. implicit conversion loses integer precision [-Wshorten-64-to-32]
2. Fix type conversions warnings for `lrint()`, which returns a `long, but almost everywhere its result is stored in an `int`, which in macOS is not the same
3. Fix deprecation warnings for `endl` vs. `Qt::endl, also seen in Windows builds in QtCreator and MSVC as clazy warnings
4. Fix more type conversion warnings
5. Disabling warnings in 3rd party code (along with reverting some fixed don to those earlier)
6. Some includes cleanup
7. Fix warnings reg. deprecated stuff, these are not Mac specific, but also seen on Windows in QtCreator as clazy warnings, marking the few remaining ToDos

+++ audio/drivers/pa.cpp
@@ -316,7 +316,7 @@ void Portaudio::putEvent(const NPlayEvent& e, unsigned framePos)
                   // Catch CTRL_PROGRAM and let other ME_CONTROLLER events to go
                   if (e.dataA() == CTRL_PROGRAM) {
                         // Convert CTRL_PROGRAM event to ME_PROGRAM
-                        long msg = Pm_Message(ME_PROGRAM | chan, less128(e.dataB()), 0);
+                        int msg = Pm_Message(ME_PROGRAM | chan, less128(e.dataB()), 0);
                         PmError error = Pm_WriteShort(portMidiDriver->getOutputStream(), seq->getCurrentMillisecondTimestampWithLatency(framePos), msg);
                         if (error != pmNoError) {
                               qDebug("Portaudio: error %d", error);
@@ -327,7 +327,7 @@ void Portaudio::putEvent(const NPlayEvent& e, unsigned framePos)
                   // fall through
             case ME_PITCHBEND:
                   {
-                  long msg = Pm_Message(e.type() | chan, less128(e.dataA()), less128(e.dataB()));
+                  int msg = Pm_Message(e.type() | chan, less128(e.dataA()), less128(e.dataB()));
                   PmError error = Pm_WriteShort(portMidiDriver->getOutputStream(), seq->getCurrentMillisecondTimestampWithLatency(framePos), msg);
                   if (error != pmNoError) {
                         qDebug("Portaudio: error %d", error);
@@ -339,7 +339,7 @@ void Portaudio::putEvent(const NPlayEvent& e, unsigned framePos)
             case ME_PROGRAM:
             case ME_AFTERTOUCH:
                   {
-                  long msg = Pm_Message(e.type() | chan, less128(e.dataA()), 0);
+                  int msg = Pm_Message(e.type() | chan, less128(e.dataA()), 0);
                   PmError error = Pm_WriteShort(portMidiDriver->getOutputStream(), seq->getCurrentMillisecondTimestampWithLatency(framePos), msg);
                   if (error != pmNoError) {
                         qDebug("Portaudio: error %d", error);
