+++ libmscore/element.cpp
@@ -2030,12 +2030,12 @@ QRectF Element::drag(EditData& ed)
       qreal _spatium = spatium();
       if (ed.hRaster) {
             qreal hRaster = _spatium / MScore::hRaster();
-            int n = lrint(x / hRaster);
+            int n = (int)lrint(x / hRaster);
             x = hRaster * n;
             }
       if (ed.vRaster) {
             qreal vRaster = _spatium / MScore::vRaster();
-            int n = lrint(y / vRaster);
+            int n = (int)lrint(y / vRaster);
             y = vRaster * n;
             }
